+++ src/webviews/components/issue/view-issue-screen/JiraIssuePage.tsx
@@ -689,6 +689,7 @@ export default class JiraIssuePage extends AbstractIssueEditorPage<Emit, Accept,
                     isAtlaskitEditorEnabled={this.state.isAtlaskitEditorEnabled}
                     onIssueUpdate={this.handleChildIssueUpdate}
                     mentionProvider={this.mentionProvider}
+                    handleEditorFocus={this.handleEditorFocus}
                 />
                 {this.advancedMain()}
                 {this.state.fields['comment'] && (
@@ -737,6 +738,7 @@ export default class JiraIssuePage extends AbstractIssueEditorPage<Emit, Accept,
                                     isEditingComment={this.state.isEditingComment}
                                     onEditingCommentChange={this.handleCommentEditingChange}
                                     mentionProvider={this.mentionProvider}
+                                    handleEditorFocus={this.handleEditorFocus}
                                 />
                             </div>
                         )}