+++ src/webviews/components/issue/view-issue-screen/mainpanel/IssueCommentComponent.tsx
@@ -36,6 +36,7 @@ export type IssueCommentComponentProps = {
     onEditingCommentChange: (editing: boolean) => void;
     isAtlaskitEditorEnabled?: boolean;
     mentionProvider: AtlascodeMentionProvider;
+    handleEditorFocus: (isFocused: boolean) => void;
 };
 const CommentComponent: React.FC<{
     siteDetails: DetailedSiteInfo;
@@ -47,6 +48,7 @@ const CommentComponent: React.FC<{
     isServiceDeskProject?: boolean;
     isAtlaskitEditorEnabled?: boolean;
     mentionProvider: AtlascodeMentionProvider;
+    handleEditorFocus: (isFocused: boolean) => void;
 }> = ({
     siteDetails,
     comment,
@@ -57,6 +59,7 @@ const CommentComponent: React.FC<{
     isServiceDeskProject,
     isAtlaskitEditorEnabled,
     mentionProvider,
+    handleEditorFocus,
 }) => {
     const { openEditor, closeEditor, isEditorActive } = useEditorState();
     const editorId = `edit-comment-${comment.id}` as const;
@@ -149,6 +152,8 @@ const CommentComponent: React.FC<{
                                     setCommentText(content);
                                 }}
                                 mentionProvider={Promise.resolve(mentionProvider)}
+                                onFocus={() => handleEditorFocus(true)}
+                                onBlur={() => handleEditorFocus(false)}
                             />
                         ) : (
                             <JiraIssueTextAreaEditor
@@ -173,6 +178,8 @@ const CommentComponent: React.FC<{
                                 }}
                                 fetchUsers={fetchUsers}
                                 isServiceDeskProject={isServiceDeskProject}
+                                onEditorFocus={() => handleEditorFocus(true)}
+                                onEditorBlur={() => handleEditorFocus(false)}
                             />
                         )
                     ) : isAtlaskitEditorEnabled ? (
@@ -201,6 +208,7 @@ const AddCommentComponent: React.FC<{
     isEditing: boolean;
     setIsEditing: (editing: boolean) => void;
     mentionProvider: AtlascodeMentionProvider;
+    handleEditorFocus: (isFocused: boolean) => void;
 }> = ({
     fetchUsers,
     user,
@@ -212,6 +220,7 @@ const AddCommentComponent: React.FC<{
     isEditing,
     setIsEditing,
     mentionProvider,
+    handleEditorFocus,
 }) => {
     const { openEditor, closeEditor } = useEditorState();
 
@@ -293,6 +302,8 @@ const AddCommentComponent: React.FC<{
                                 setCommentText(content);
                             }}
                             mentionProvider={Promise.resolve(mentionProvider)}
+                            onFocus={() => handleEditorFocus(true)}
+                            onBlur={() => handleEditorFocus(false)}
                         />
                     </Box>
                 ) : (
@@ -315,7 +326,11 @@ const AddCommentComponent: React.FC<{
                             setCommentText('');
                             closeEditorHandler();
                         }}
-                        onEditorFocus={openEditorHandler}
+                        onEditorFocus={() => {
+                            openEditorHandler();
+                            handleEditorFocus(true);
+                        }}
+                        onEditorBlur={() => handleEditorFocus(false)}
                         fetchUsers={fetchUsers}
                         isServiceDeskProject={isServiceDeskProject}
                     />
@@ -340,6 +355,7 @@ export const IssueCommentComponent: React.FC<IssueCommentComponentProps> = ({
     onEditingCommentChange,
     isAtlaskitEditorEnabled,
     mentionProvider,
+    handleEditorFocus,
 }) => {
     return (
         <Box
@@ -357,6 +373,7 @@ export const IssueCommentComponent: React.FC<IssueCommentComponentProps> = ({
                 isEditing={isEditingComment}
                 setIsEditing={onEditingCommentChange}
                 mentionProvider={mentionProvider}
+                handleEditorFocus={handleEditorFocus}
             />
             {comments
                 .sort((a, b) => (a.created > b.created ? -1 : 1))
@@ -372,6 +389,7 @@ export const IssueCommentComponent: React.FC<IssueCommentComponentProps> = ({
                         isServiceDeskProject={isServiceDeskProject}
                         isAtlaskitEditorEnabled={isAtlaskitEditorEnabled}
                         mentionProvider={mentionProvider}
+                        handleEditorFocus={handleEditorFocus}
                     />
                 ))}
         </Box>