+++ src/webviews/components/issue/view-issue-screen/mainpanel/IssueMainPanel.tsx
@@ -43,6 +43,7 @@ type Props = {
     onIssueUpdate?: (issueKey: string, fieldKey: string, newValue: any) => void;
     isAtlaskitEditorEnabled?: boolean;
     mentionProvider: AtlascodeMentionProvider;
+    handleEditorFocus: (isFocused: boolean) => void;
 };
 
 const IssueMainPanel: React.FC<Props> = ({
@@ -66,6 +67,7 @@ const IssueMainPanel: React.FC<Props> = ({
     onIssueUpdate,
     isAtlaskitEditorEnabled,
     mentionProvider,
+    handleEditorFocus,
 }) => {
     const attachments = fields['attachment'] && fieldValues['attachment'] ? fieldValues['attachment'] : undefined;
     const subtasks =
@@ -247,6 +249,8 @@ const IssueMainPanel: React.FC<Props> = ({
                                     setDescriptionText(content);
                                 }}
                                 mentionProvider={Promise.resolve(mentionProvider)}
+                                onFocus={() => handleEditorFocus(true)}
+                                onBlur={() => handleEditorFocus(false)}
                             />
                         ) : (
                             <JiraIssueTextAreaEditor
@@ -265,6 +269,8 @@ const IssueMainPanel: React.FC<Props> = ({
                                 fetchUsers={fetchUsers}
                                 isDescription
                                 saving={loadingField === 'description'}
+                                onEditorFocus={() => handleEditorFocus(true)}
+                                onEditorBlur={() => handleEditorFocus(false)}
                             />
                         )
                     ) : (