+++ src/webviews/jiraIssueWebview.ts
@@ -37,6 +37,7 @@ import {
     isDeleteByIDAction,
     isDeleteWorklog,
     isGetImage,
+    isHandleEditorFocus,
     isIssueComment,
     isIssueDeleteComment,
     isOpenStartWorkPageAction,
@@ -1562,6 +1563,13 @@ export class JiraIssueWebview
                             type: 'historyUpdate',
                             history: [],
                         });
+                    }
+                    break;
+                }
+                case 'handleEditorFocus': {
+                    if (isHandleEditorFocus(msg)) {
+                        handled = true;
+                        Container.setIsEditorFocused(msg.isFocused);
                     }
                     break;
                 }